% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_interactive.R
\name{geom_image_interactive}
\alias{geom_image_interactive}
\title{Create interactive image of ggimage}
\usage{
geom_image_interactive(...)
}
\arguments{
\item{...}{see also the parameters of \code{geom_image()} of \code{ggimage}}
}
\description{
The geometry is based on \code{geom_image()}.
See the documentation for those functions for more details.
}
