% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_interactive.R
\name{geom_phylopic_interactive}
\alias{geom_phylopic_interactive}
\title{Create interactive phylopic of ggimage}
\usage{
geom_phylopic_interactive(...)
}
\arguments{
\item{...}{see also the parameters of \code{geom_phylopic()} of \code{ggimage}}
}
\description{
The geometry is based on \code{geom_phylopic()}.
See the documentation for those functions for more details.
}
