% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_interactive.R
\name{geom_point2_interactive}
\alias{geom_point2_interactive}
\title{Create interactive point of ggtree}
\usage{
geom_point2_interactive(...)
}
\arguments{
\item{...}{see also the parameters of \code{geom_point2()} of \code{ggtree}}
}
\description{
The geometry is based on \code{geom_point2()}.
See the documentation for those functions for more details.
}
