% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_interactive.R
\name{geom_text2_interactive}
\alias{geom_text2_interactive}
\title{Create interactive text of ggtree}
\usage{
geom_text2_interactive(...)
}
\arguments{
\item{...}{see also the parameters of \code{geom_text2()} of \code{ggtree}}
}
\description{
The geometry is based on \code{geom_text2()}.
See the documentation for those functions for more details.
}
